Added workaround for seek error
When seeking we would like to go to an exact frame, but av has issues in doing so, and we can't iterate from the closest keyframe as av won't tell us what keyframe it seeked to. So, as this will be replaced anyway, it just goes to the closest keyframe.